requirements.txt
altgraph==0.17.4
attrs==25.1.0
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_2emnmf1qow/croot/certifi_1690232232440/work/certifi
charset-normalizer==3.4.1
click==8.1.7
et-xmlfile==1.1.0
exceptiongroup==1.2.2
filelock==3.13.1
fsspec==2024.3.1
h11==0.14.0
idna==3.10
importlib-metadata==6.8.0
Jinja2==3.1.3
macholib==1.16.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.2
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==23.2
pandas==1.5.3
Pillow==9.4.0
PyInstaller==3.6
pyinstaller-hooks-contrib==2023.9
pypng==0.20220715.0
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
pyqt5-tools==5.15.9.3.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
pyuic5-tool==0.0.1
qrcode==7.4.2
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
requests==2.32.3
selenium==4.27.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sympy==1.12
torch==2.2.1
torchvision==0.17.1
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.10.0
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
XlsxWriter==3.0.9
zipp==3.17.0
